:art: turn the list of drivers into a column
There are a finite number of drivers that can be compiled into `drmemd`
so just build a column of rows. The body of the window is scrollable and
when a ListView was used, you could scroll the window when the
ListView took up all the space because the drag gesture was being
consumed by the ListView.