:recycle: reorganize GraphQL logic in SheetRunner
The subscription client was moved out of the `DeviceWidget` and into the
`SheetRunner`. This was done so that only one WebSocket is opened
(previously each DeviceRow would open a new WebSocket.)

We now handle some error status and display it to the user.

The GraphQL requests are now sequenced. First a query is sent to get the
device's information. When that completes, and the info is stored away, it
sets up the subscription for reading the device.