:zap: optimize displaying device data
Prior to this commit, each row in the `SheetRunner` widget monitored the
device and when there was an update, the entire row re-rendered itself.
I noticed it wasn't keeping up well with a 1 Hz rate and the profiler showed
it was spending more CPU than I would like. One a screenful of devices, this
would be very noticeable.

This commit makes the widget that shows the actual value is in charge of
subscribing to device changes. Only it has to re-render -- not the whole
row.