:bug: add `sec-websocket-protocol` header for subs
Subscriptions use WebSockets to receive streamed data. The WebSocket RFC
poorly defines how `sec-websocket-protocol` should be handled. We know
that, in the case of Chrome Browsers, if the client doesn't send the
header, Chrome doesn;t want the reply to include it. This commit adds
header to make Chrome happy.