:bug: fix highlighting
Add a `Key` to the buttons so Flutter can keep track of which was
pressed. Prior to this, when you pressed the edit button the add and
delete buttons would appear and the add button would receive the
ending of the hightlight animation.

Now the edit button finishes its animation.