:construction: massive changes
This commit wraps up a large scale reorganization of the code. One effort
was to bring in the DrMem widget package instead of doing the GraphQL
ourselves. But, while doing it, I also modified some of the functionality.

This code is not ready for production and needs a lot of testing. Also, the
node detection needs to be enhanced so detected nodes are added to our
persistent storage. Also, all device access is being done to a hardcoded
node called "rpi4".